requirements.txt
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
starlette>=0.40.0
uvicorn>=0.30.0
fastapi>=0.115.0
pyyaml>=6.0
